+++ ultrasonic.h
@@ -3,6 +3,6 @@
 void init_ultrasonic(void);
 void PIT0_IRQHandler(void);
 void init_PIT(void);
-float getUltrasonic(void);
+double getUltrasonic(void);
 //void UART0_RX_TX_IRQHandler(void);
 #endif /* ultrasonic_H_ */
